refactor: move functions, cleanup
